Dynamic scheme handler registry for plugins and clener core
